Abstracted out LearningAlgorithm.
Added the LearningAlgorithm base class. Previously, the functionality of
what is now LearningAlgorithm was bundled up with policies (so policies
had a `fit` method, for example). Now these have been teased apart,
yielding a clearer interface. The `fit` method of an instance of a
class that derives from LearningAlgorithm should return a policy.