Big commit making DAgger work in all cases.
The three cases are: fully observable dagger, partially observable
dagger on belief states, and partially observable dagger with spectral
learning.

Had to reformulate the interface to a few of the classes. Had to make
BeliefStatePolicy a class that could actually be implemented, accepting
a function which maps belief states to actions.

Had to add the `dimension` attribute to the State class, so that the
`as_vector` method could be added.