Now using complete reward matrix.
Reward is now a function of action, state and next state. Previously it
was just action and next state, but that can be fairly limiting.

Now, there is an option to omit the next state if desired when calling
get reward. In this case, it will return the expected reward from
transitioning from state s using action a.