GridWorld improvements.
Added a class `Position`, which will be used more or less everywhere.
Combines the easy manipulation of ndarrays with the ability to easily
test for inequality, and the ability to use as indices into the
WorldMap.

Also added the WorldMap class. Each GridWorld has an instance of the
WorldMap class, which stores all the positional information. The
responsibility of the GridWorld class is then providing an MDP-like
wrapper on to of the WorldMap (so handling actions and rewards).