Revamped random number generation.
There is now a dictionary in the top-level package which contains random
number generators. It should be accessed through the top-level function
``rngs``. When accessed this way, it acts like a defaultdict, so that
keys can be added simply by accessing them (though they can also be
added directly with a specified value, also by calling ``rngs``). When
using the defaultdict behaviour, if a new rng must be generated, the
seed for said rng is generated from the current default rng (``rngs[None]``).

Also a function ``set_seed`` which sets the default rng (_rngs[None]) to
a new rng using the given seed, and clears the contents of ``_rngs``,
so that all subsequent uses of any random number generator from
``_rngs`` are influenced by the new seed.